Fix-ups post-move.
