+++ state/resources_persistence_test.go
@@ -1,7 +1,7 @@
 // Copyright 2015 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package persistence
+package state
 
 import (
 	"time"
@@ -16,7 +16,7 @@ import (
 
 	"github.com/juju/juju/resource"
 	"github.com/juju/juju/resource/resourcetesting"
-	statetest "github.com/juju/juju/state/testing"
+	"github.com/juju/juju/state/statetest"
 )
 
 var _ = gc.Suite(&ResourcePersistenceSuite{})
@@ -36,10 +36,10 @@ func (s *ResourcePersistenceSuite) SetUpTest(c *gc.C) {
 }
 
 func (s *ResourcePersistenceSuite) TestListResourcesOkay(c *gc.C) {
-	expected, docs := newResources(c, "a-service", "spam", "eggs")
+	expected, docs := newPersistenceResources(c, "a-service", "spam", "eggs")
 	expected.CharmStoreResources[1].Revision += 1
 	docs[3].Revision += 1
-	unitRes, doc := newUnitResource(c, "a-service", "a-service/0", "something")
+	unitRes, doc := newPersistenceUnitResource(c, "a-service", "a-service/0", "something")
 	expected.UnitResources = []resource.UnitResources{{
 		Tag: names.NewUnitTag("a-service/0"),
 		Resources: []resource.Resource{
@@ -77,7 +77,7 @@ func (s *ResourcePersistenceSuite) TestListResourcesNoResources(c *gc.C) {
 }
 
 func (s *ResourcePersistenceSuite) TestListResourcesIgnorePending(c *gc.C) {
-	expected, docs := newResources(c, "a-service", "spam", "eggs")
+	expected, docs := newPersistenceResources(c, "a-service", "spam", "eggs")
 	expected.Resources = expected.Resources[:1]
 	docs[2].PendingID = "some-unique-ID-001"
 	s.base.ReturnAll = docs
@@ -112,7 +112,7 @@ func (s *ResourcePersistenceSuite) TestListResourcesBaseError(c *gc.C) {
 }
 
 func (s *ResourcePersistenceSuite) TestListResourcesBadDoc(c *gc.C) {
-	_, docs := newResources(c, "a-service", "spam", "eggs")
+	_, docs := newPersistenceResources(c, "a-service", "spam", "eggs")
 	docs[0].Timestamp = time.Time{}
 	s.base.ReturnAll = docs
 
@@ -132,7 +132,7 @@ func (s *ResourcePersistenceSuite) TestListPendingResourcesOkay(c *gc.C) {
 	var expected []resource.Resource
 	var docs []resourceDoc
 	for _, name := range []string{"spam", "ham"} {
-		res, doc := newResource(c, "a-service", name)
+		res, doc := newPersistenceResource(c, "a-service", name)
 		expected = append(expected, res.Resource)
 		docs = append(docs, doc)
 	}
@@ -155,7 +155,7 @@ func (s *ResourcePersistenceSuite) TestListPendingResourcesOkay(c *gc.C) {
 }
 
 func (s *ResourcePersistenceSuite) TestGetResourceOkay(c *gc.C) {
-	expected, doc := newResource(c, "a-service", "spam")
+	expected, doc := newPersistenceResource(c, "a-service", "spam")
 	unitDoc := doc // a copy
 	unitDoc.ID = doc.ID + "#unit-a-service/0"
 	unitDoc.UnitID = "a-service/0"
@@ -180,7 +180,7 @@ func (s *ResourcePersistenceSuite) TestGetResourceOkay(c *gc.C) {
 }
 
 func (s *ResourcePersistenceSuite) TestStageResourceOkay(c *gc.C) {
-	res, doc := newResource(c, "a-service", "spam")
+	res, doc := newPersistenceResource(c, "a-service", "spam")
 	doc.DocID += "#staged"
 	p := NewResourcePersistence(s.base)
 	ignoredErr := errors.New("<never reached>")
@@ -204,7 +204,7 @@ func (s *ResourcePersistenceSuite) TestStageResourceOkay(c *gc.C) {
 }
 
 func (s *ResourcePersistenceSuite) TestStageResourceMissingStoragePath(c *gc.C) {
-	res, _ := newResource(c, "a-service", "spam")
+	res, _ := newPersistenceResource(c, "a-service", "spam")
 	p := NewResourcePersistence(s.base)
 
 	_, err := p.StageResource(res.Resource, "")
@@ -214,7 +214,7 @@ func (s *ResourcePersistenceSuite) TestStageResourceMissingStoragePath(c *gc.C) 
 }
 
 func (s *ResourcePersistenceSuite) TestStageResourceBadResource(c *gc.C) {
-	res, _ := newResource(c, "a-service", "spam")
+	res, _ := newPersistenceResource(c, "a-service", "spam")
 	res.Resource.Timestamp = time.Time{}
 	p := NewResourcePersistence(s.base)
 
@@ -228,7 +228,7 @@ func (s *ResourcePersistenceSuite) TestStageResourceBadResource(c *gc.C) {
 
 func (s *ResourcePersistenceSuite) TestSetResourceOkay(c *gc.C) {
 	servicename := "a-service"
-	res, doc := newResource(c, servicename, "spam")
+	res, doc := newPersistenceResource(c, servicename, "spam")
 	s.base.ReturnOne = doc
 	p := NewResourcePersistence(s.base)
 	ignoredErr := errors.New("<never reached>")
@@ -252,7 +252,7 @@ func (s *ResourcePersistenceSuite) TestSetResourceOkay(c *gc.C) {
 
 func (s *ResourcePersistenceSuite) TestSetResourceNotFound(c *gc.C) {
 	servicename := "a-service"
-	res, doc := newResource(c, servicename, "spam")
+	res, doc := newPersistenceResource(c, servicename, "spam")
 	s.base.ReturnOne = doc
 	expected := doc // a copy
 	expected.StoragePath = ""
@@ -280,7 +280,7 @@ func (s *ResourcePersistenceSuite) TestSetResourceNotFound(c *gc.C) {
 func (s *ResourcePersistenceSuite) TestSetCharmStoreResourceOkay(c *gc.C) {
 	lastPolled := time.Now().UTC()
 	servicename := "a-service"
-	res, doc := newResource(c, servicename, "spam")
+	res, doc := newPersistenceResource(c, servicename, "spam")
 	expected := doc // a copy
 	expected.DocID += "#charmstore"
 	expected.Username = ""
@@ -309,7 +309,7 @@ func (s *ResourcePersistenceSuite) TestSetCharmStoreResourceOkay(c *gc.C) {
 func (s *ResourcePersistenceSuite) TestSetUnitResourceOkay(c *gc.C) {
 	servicename := "a-service"
 	unitname := "a-service/0"
-	res, doc := newUnitResource(c, servicename, unitname, "eggs")
+	res, doc := newPersistenceUnitResource(c, servicename, unitname, "eggs")
 	s.base.ReturnOne = doc
 	p := NewResourcePersistence(s.base)
 	ignoredErr := errors.New("<never reached>")
@@ -330,7 +330,7 @@ func (s *ResourcePersistenceSuite) TestSetUnitResourceOkay(c *gc.C) {
 func (s *ResourcePersistenceSuite) TestSetUnitResourceNotFound(c *gc.C) {
 	servicename := "a-service"
 	unitname := "a-service/0"
-	res, _ := newUnitResource(c, servicename, unitname, "eggs")
+	res, _ := newPersistenceUnitResource(c, servicename, unitname, "eggs")
 	p := NewResourcePersistence(s.base)
 	notFound := errors.NewNotFound(nil, "")
 	s.stub.SetErrors(notFound)
@@ -343,7 +343,7 @@ func (s *ResourcePersistenceSuite) TestSetUnitResourceNotFound(c *gc.C) {
 }
 
 func (s *ResourcePersistenceSuite) TestSetUnitResourceExists(c *gc.C) {
-	res, doc := newUnitResource(c, "a-service", "a-service/0", "spam")
+	res, doc := newPersistenceUnitResource(c, "a-service", "a-service/0", "spam")
 	s.base.ReturnOne = doc
 	p := NewResourcePersistence(s.base)
 	ignoredErr := errors.New("<never reached>")
@@ -373,7 +373,7 @@ func (s *ResourcePersistenceSuite) TestSetUnitResourceExists(c *gc.C) {
 }
 
 func (s *ResourcePersistenceSuite) TestSetUnitResourceBadResource(c *gc.C) {
-	res, doc := newUnitResource(c, "a-service", "a-service/0", "spam")
+	res, doc := newPersistenceUnitResource(c, "a-service", "a-service/0", "spam")
 	s.base.ReturnOne = doc
 	res.Timestamp = time.Time{}
 	p := NewResourcePersistence(s.base)
@@ -388,7 +388,7 @@ func (s *ResourcePersistenceSuite) TestSetUnitResourceBadResource(c *gc.C) {
 
 func (s *ResourcePersistenceSuite) TestNewResourcePendingResourceOpsExists(c *gc.C) {
 	pendingID := "some-unique-ID-001"
-	stored, expected := newResource(c, "a-service", "spam")
+	stored, expected := newPersistenceResource(c, "a-service", "spam")
 	stored.PendingID = pendingID
 	doc := expected // a copy
 	doc.DocID = pendingResourceID(stored.ID, pendingID)
@@ -421,7 +421,7 @@ func (s *ResourcePersistenceSuite) TestNewResourcePendingResourceOpsExists(c *gc
 
 func (s *ResourcePersistenceSuite) TestNewResourcePendingResourceOpsNotFound(c *gc.C) {
 	pendingID := "some-unique-ID-001"
-	stored, expected := newResource(c, "a-service", "spam")
+	stored, expected := newPersistenceResource(c, "a-service", "spam")
 	stored.PendingID = pendingID
 	doc := expected // a copy
 	doc.DocID = pendingResourceID(stored.ID, pendingID)
@@ -449,11 +449,11 @@ func (s *ResourcePersistenceSuite) TestNewResourcePendingResourceOpsNotFound(c *
 	}})
 }
 
-func newResources(c *gc.C, serviceID string, names ...string) (resource.ServiceResources, []resourceDoc) {
+func newPersistenceResources(c *gc.C, serviceID string, names ...string) (resource.ServiceResources, []resourceDoc) {
 	var svcResources resource.ServiceResources
 	var docs []resourceDoc
 	for _, name := range names {
-		res, doc := newResource(c, serviceID, name)
+		res, doc := newPersistenceResource(c, serviceID, name)
 		svcResources.Resources = append(svcResources.Resources, res.Resource)
 		svcResources.CharmStoreResources = append(svcResources.CharmStoreResources, res.Resource.Resource)
 		docs = append(docs, doc)
@@ -468,14 +468,14 @@ func newResources(c *gc.C, serviceID string, names ...string) (resource.ServiceR
 	return svcResources, docs
 }
 
-func newUnitResource(c *gc.C, serviceID, unitID, name string) (resource.Resource, resourceDoc) {
-	res, doc := newResource(c, serviceID, name)
+func newPersistenceUnitResource(c *gc.C, serviceID, unitID, name string) (resource.Resource, resourceDoc) {
+	res, doc := newPersistenceResource(c, serviceID, name)
 	doc.DocID += "#unit-" + unitID
 	doc.UnitID = unitID
 	return res.Resource, doc
 }
 
-func newResource(c *gc.C, serviceID, name string) (storedResource, resourceDoc) {
+func newPersistenceResource(c *gc.C, serviceID, name string) (storedResource, resourceDoc) {
 	content := name
 	opened := resourcetesting.NewResource(c, nil, name, serviceID, content)
 	res := opened.Resource
