+++ state/statetest/persistence_stubs.go
@@ -1,7 +1,7 @@
 // Copyright 2015 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package testing
+package statetest
 
 import (
 	"reflect"
